requirements.txt
Pillow
flask-restx
flask-cors
flask
streamlit
requests